requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==0.27.0  # Ensure compatibility based on prior issues
pytest==7.1.2
